CORE-79.
Remove Yarrow.java, as it may have come from the Freenet codebase, which is GPL.

As it is only used internally for generating GUIDs, this is an easy fix, but does make the codebase JDK 1.4 reliant.  I don't think this is an issue, as Sun itself doesn't support 1.3 anymore.

If you relied on Yarrow.java, then you will need to find a different implementation